agents/sourcery/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pyyaml>=6.0
python-dotenv>=1.0.0
google-adk>=1.1.0
google-genai>=1.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
google-cloud-storage>=2.14.0
